@@ -1,69 +1,54 @@
 {
-  "name": "client",
-  "version": "0.1.0",
+  "name": "soulforged-client",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build-secure-components": "./tools/build-secure-components.sh",
-    "build-secure-components:watch": "./tools/build-secure-components.sh -watch",
-    "build": "yarn vue-cli-service build",
-    "build:watch": "yarn build --watch",
-    "build-all": "npm-run-all --parallel build build-secure-components",
-    "build-all:watch": "npm-run-all --parallel build:watch build-secure-components build-secure-components:watch",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook -o .storybook/dist",
-    "test-storyshots": "npm-run-all build-storybook test-storyshots:run",
-    "test-storyshots:update": "npm-run-all build-storybook test-storyshots:run-update",
-    "test-storyshots:run": "jest storyshots.runner.js --testRegex='.'",
-    "test-storyshots:run-update": "jest storyshots.runner.js --testRegex='.' -u"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test:unit": "vitest",
+    "lint": "eslint . --fix",
+    "format": "prettier --write src/",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "dependencies": {
-    "core-js": "^3.6.5",
-    "howler": "^2.2.3",
-    "jest-vue-preprocessor": "^1.7.1",
+    "howler": "^2.2.4",
     "lodash": "^4.17.21",
     "random-seed": "^0.3.0",
-    "rxjs": "^6.6.3",
-    "typescript": "^4.7.3",
-    "vue": "^2.7.14",
-    "vue-observe-visibility": "^1.0.0",
-    "vue-router": "^3.2.0",
-    "vue-rx": "^6.2.0"
+    "storybook": "^8.6.7",
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@storybook/addon-actions": "^6.5.14",
-    "@storybook/addon-essentials": "^6.5.14",
-    "@storybook/addon-links": "^6.5.14",
-    "@storybook/addon-storyshots": "^6.5.14",
-    "@storybook/addon-storyshots-puppeteer": "^6.5.14",
-    "@storybook/preset-scss": "^1.0.3",
-    "@storybook/vue": "6.5.14",
-    "@vue/cli-plugin-babel": "~5.0.8",
-    "@vue/cli-plugin-router": "~5.0.8",
-    "@vue/cli-service": "~5.0.8",
-    "babel-loader": "^9.1.3",
-    "css-loader": "^7.1.2",
-    "cytoscape-cose-bilkent": "^4.1.0",
-    "cytoscape-dagre": "^2.3.2",
-    "cytoscape-klay": "^3.1.4",
-    "git-describe": "^4.0.4",
-    "jest-environment-jsdom": "^28.1.1",
-    "npm-run-all": "^4.1.5",
-    "puppeteer": "^7.0.1",
-    "sass": "^1.26.5",
-    "sass-loader": "^10.1.1",
-    "storybook": "^6.5.14",
-    "style-loader": "^2.0.0",
-    "vue-cytoscape": "^1.0.8",
-    "vue-loader": "^17.4.2",
-    "vue-template-compiler": "2.7.16"
+    "@chromatic-com/storybook": "^3",
+    "@eslint/js": "^9.21.0",
+    "@storybook/addon-essentials": "^8.6.7",
+    "@storybook/addon-interactions": "^8.6.7",
+    "@storybook/addon-onboarding": "^8.6.7",
+    "@storybook/blocks": "^8.6.7",
+    "@storybook/test": "^8.6.7",
+    "@storybook/vue3": "^8.6.7",
+    "@storybook/vue3-vite": "^8.6.7",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "@vitest/eslint-plugin": "^1.1.36",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "@vue/test-utils": "^2.4.6",
+    "eslint": "^9.21.0",
+    "eslint-plugin-storybook": "^0.11.6",
+    "eslint-plugin-vue": "~10.0.0",
+    "globals": "^16.0.0",
+    "jsdom": "^26.0.0",
+    "prettier": "3.5.3",
+    "sass-embedded": "^1.86.0",
+    "vite": "^6.2.1",
+    "vite-plugin-vue-devtools": "^7.7.2",
+    "vitest": "^3.0.8"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ],
-  "engines": {
-    "node": ">=22.0.0"
+  "eslintConfig": {
+    "extends": [
+      "plugin:storybook/recommended"
+    ]
   }
 }
