@@ -1,32 +1,31 @@
-import sampleMapImage from "./assets/map.jpg";
-import "../../common/utils/static.js";
-import Vue from "vue";
-import { Rx } from "@/rx.js";
-import flatten from "lodash/flatten.js";
+import sampleMapImage from './assets/map.jpg'
+import '../../common/utils/static.js'
+import { Rx } from '../../src/rx.js'
+import flatten from 'lodash/flatten.js'
 
 const CONTROL_TYPES = {
-  string: "text",
-  boolean: "boolean",
-};
-const pluginParamsMap = {};
+  string: 'text',
+  boolean: 'boolean',
+}
+const pluginParamsMap = {}
 const pluginParams = flatten(
-  window.allPlugins.map((def) => {
+  (window.allPlugins || []).map((def) => {
     if (!def.enabled || def.error) {
-      return [];
+      return []
     }
-    const key = `Plugin ${def.name}: Enabled`;
-    pluginParamsMap[key] = [def.id, "enabled"];
+    const key = `Plugin ${def.name}: Enabled`
+    pluginParamsMap[key] = [def.id, 'enabled']
     return [
       {
         key,
         value: false,
         spec: {
-          control: { type: "boolean" },
+          control: { type: 'boolean' },
         },
       },
       ...(def.settings || []).map((setting) => {
-        const key = `Plugin ${def.name}: ${setting.name}`;
-        pluginParamsMap[key] = [def.id, setting.id];
+        const key = `Plugin ${def.name}: ${setting.name}`
+        pluginParamsMap[key] = [def.id, setting.id]
         return {
           key,
           value: null,
@@ -34,154 +33,152 @@ const pluginParams = flatten(
             control: { type: CONTROL_TYPES[setting.type] },
             if: { arg: `Plugin ${def.name}: Enabled`, truthy: true },
           },
-        };
+        }
       }),
-    ];
-  })
-);
+    ]
+  }),
+)
 
 global.loremIpsum =
-  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
 
-global.IDS = {};
-global.idToIdentifier = {};
-global.entitiesMocks = {};
+global.IDS = {}
+global.idToIdentifier = {}
+global.entitiesMocks = {}
 
-let ids = 90000;
-let idsGenerated = false;
+let ids = 90000
+let idsGenerated = false
 global.mockEntity = (identifier, definitionFn) => {
-  const id = ids++;
+  const id = ids++
   onIdsGenerated(() => {
     entitiesMocks[identifier] = {
       ...entitiesMocks[identifier],
       ...definitionFn(),
-    };
-  });
-  idToIdentifier[id] = identifier;
+    }
+  })
+  idToIdentifier[id] = identifier
   return (entitiesMocks[identifier] = {
     ...(entitiesMocks[identifier] || {}),
     id,
-  });
-};
-const entityGenerators = [];
+  })
+}
+const entityGenerators = []
 global.onIdsGenerated = (callback) => {
   if (!idsGenerated) {
-    entityGenerators.push(callback);
+    entityGenerators.push(callback)
   } else {
-    callback();
+    callback()
   }
-};
+}
 global.allIdsGenerated = () => {
-  console.log(`Generating ${entityGenerators.length} entities payloads...`);
-  entityGenerators.forEach((callback) => callback());
-  idsGenerated = true;
-  console.log("Payloads ready");
-};
+  console.log(`Generating ${entityGenerators.length} entities payloads...`)
+  entityGenerators.forEach((callback) => callback())
+  idsGenerated = true
+  console.log('Payloads ready')
+}
 
 global.mockEntityId = (identifier) => {
   if (!entitiesMocks[identifier]) {
     throw new Error(
-      `Invalid entityId ${identifier}, available: ${Object.keys(
-        entitiesMocks
-      ).join(", ")}`
-    );
+      `Invalid entityId ${identifier}, available: ${Object.keys(entitiesMocks).join(', ')}`,
+    )
   }
-  return entitiesMocks[identifier].id;
-};
-let entityStreams = {};
+  return entitiesMocks[identifier].id
+}
+let entityStreams = {}
 const stream = (id) => {
   if (!entityStreams[id]) {
-    entityStreams[id] = new Rx.ReplaySubject(1);
-    entityStreams[id].next(entitiesMocks[idToIdentifier[id]]);
+    entityStreams[id] = new Rx.ReplaySubject(1)
+    entityStreams[id].next(entitiesMocks[idToIdentifier[id]])
   }
-  return entityStreams[id];
-};
+  return entityStreams[id]
+}
 global.mockComponent = (
   definition,
   onMounted = () => {},
   onCreate = () => {},
-  onDestroyed = () => {}
+  onDestroyed = () => {},
 ) => {
   const storyFunction = (args) => {
-    const updatedPluginSettings = {};
+    const updatedPluginSettings = {}
     Object.keys(args).forEach((key) => {
       if (pluginParamsMap[key]) {
-        const [plugin, setting] = pluginParamsMap[key];
+        const [plugin, setting] = pluginParamsMap[key]
         if (!updatedPluginSettings[plugin]) {
-          updatedPluginSettings[plugin] = {};
+          updatedPluginSettings[plugin] = {}
         }
-        updatedPluginSettings[plugin][setting] = args[key];
+        updatedPluginSettings[plugin][setting] = args[key]
       }
-    });
-    PluginService.mockUpdateSettingStream(updatedPluginSettings);
+    })
+    PluginService.mockUpdateSettingStream(updatedPluginSettings)
 
-    entityStreams = {};
+    entityStreams = {}
     GameService.mock({
       getLocationImgPath: () => sampleMapImage,
       getEntityStream: (id) => stream(id),
-      getRootEntityStream: () => stream(entitiesMocks["mainEntity"].id),
-      getLocationStream: () => stream(entitiesMocks["location"].id),
-      getKnowledgeBaseStream: () => stream(entitiesMocks["knowledgeBase"].id),
+      getRootEntityStream: () => stream(entitiesMocks['mainEntity'].id),
+      getLocationStream: () => stream(entitiesMocks['location'].id),
+      getKnowledgeBaseStream: () => stream(entitiesMocks['knowledgeBase'].id),
       getQuickActionsStream: () => Rx.Observable.of([]),
-    });
+    })
 
-    onCreate();
+    onCreate()
 
-    Vue.prototype.$router = {
-      push: () => {},
-      currentRoute: {
-        query: {},
-      },
-    };
-    Vue.prototype.$route = {
-      query: {},
-    };
+    // Vue.prototype.$router = {
+    //   push: () => {},
+    //   currentRoute: {
+    //     query: {},
+    //   },
+    // };
+    // Vue.prototype.$route = {
+    //   query: {},
+    // };
 
     return {
       mounted() {
-        onMounted(this);
+        onMounted(this)
       },
 
       beforeDestroy() {
-        onDestroyed(this);
+        onDestroyed(this)
       },
 
       ...definition,
-    };
-  };
+    }
+  }
 
   storyFunction.args = pluginParams.toObject(
     (def) => def.key,
-    (def) => def.value
-  );
+    (def) => def.value,
+  )
   storyFunction.argTypes = pluginParams.toObject(
     (def) => def.key,
-    (def) => def.spec
-  );
+    (def) => def.spec,
+  )
 
-  return storyFunction;
-};
+  return storyFunction
+}
 
 export const modifyEntity = (entityId, object) => {
-  const id = entitiesMocks[entityId]?.id || entityId;
+  const id = entitiesMocks[entityId]?.id || entityId
   stream(id)
     .first()
     .subscribe((current) => {
-      console.log("UPDATED", id, object);
+      console.log('UPDATED', id, object)
       stream(id).next({
         ...current,
         ...object,
-      });
-    });
-};
+      })
+    })
+}
 
 export const cloneEntityPayloadId = (entityId, extras = {}) => {
-  const id = ids++;
+  const id = ids++
   entitiesMocks[id] = {
     ...entitiesMocks[entityId],
     id,
     ...extras,
-  };
-  idToIdentifier[id] = id;
-  return id;
-};
+  }
+  idToIdentifier[id] = id
+  return id
+}
