@@ -1,62 +1,225 @@
-import Vue from "vue";
+import Backgrounds from './src/components/layouts/Backgrounds.vue'
+import BorderRound from './src/components/layouts/BorderRound.vue'
+import Container from './src/components/layouts/Container.vue'
+import DynamicLayout from './src/components/layouts/DynamicLayout.vue'
+import Horizontal from './src/components/layouts/Horizontal.vue'
+import HorizontalCenter from './src/components/layouts/HorizontalCenter.vue'
+import HorizontalFill from './src/components/layouts/HorizontalFill.vue'
+import HorizontalWrap from './src/components/layouts/HorizontalWrap.vue'
+import PageLayout from './src/components/layouts/PageLayout.vue'
+import Spaced from './src/components/layouts/Spaced.vue'
+import Vertical from './src/components/layouts/Vertical.vue'
 
-const components = {};
-const allPlugins = {};
+// interface
+import AppStyles from './src/components/interface/AppStyles.vue'
+import AutoSizer from './src/components/interface/AutoSizer.vue'
+import Button from './src/components/interface/Button.vue'
+import Checkbox from './src/components/interface/Checkbox.vue'
+import CloseButton from './src/components/interface/CloseButton.vue'
+import Collapsible from './src/components/interface/Collapsible.vue'
+import Countdown from './src/components/interface/Countdown.vue'
+import Description from './src/components/interface/Description.vue'
+import EditButton from './src/components/interface/EditButton.vue'
+import ExplanationIndicator from './src/components/interface/ExplanationIndicator.vue'
+import Header from './src/components/interface/Header.vue'
+import Help from './src/components/interface/Help.vue'
+import Icon from './src/components/interface/Icon.vue'
+import Input from './src/components/interface/Input.vue'
+import LabeledValue from './src/components/interface/LabeledValue.vue'
+import ListItem from './src/components/interface/ListItem.vue'
+import LoadingPlaceholder from './src/components/interface/LoadingPlaceholder.vue'
+import Modal from './src/components/interface/Modal.vue'
+import OnlyIfVisible from './src/components/interface/OnlyIfVisible.vue'
+import OptionSelector from './src/components/interface/OptionSelector.vue'
+import PluginInjection from './src/components/interface/PluginInjection.vue'
+import Plused from './src/components/interface/Plused.vue'
+import ProgressBar from './src/components/interface/ProgressBar.vue'
+import Radio from './src/components/interface/Radio.vue'
+import ReportButton from './src/components/interface/ReportButton.vue'
+import Scrollbar from './src/components/interface/Scrollbar.vue'
+import Select from './src/components/interface/Select.vue'
+import Slider from './src/components/interface/Slider.vue'
+import Spinner from './src/components/interface/Spinner.vue'
+import StarRating from './src/components/interface/StarRating.vue'
+import SubLabeledValue from './src/components/interface/SubLabeledValue.vue'
+import TextArea from './src/components/interface/TextArea.vue'
+import ToastNotifications from './src/components/interface/ToastNotifications.vue'
+import TopZIndex from './src/components/interface/TopZIndex.vue'
+import ViewButton from './src/components/interface/ViewButton.vue'
 
-if (typeof require.context === "undefined") {
-  require.context = () => ({
-    keys: () => [],
-  });
+//game
+import Actions from './src/components/game/Actions.vue'
+import APBar from './src/components/game/APBar.vue'
+import APBarCurrent from './src/components/game/APBarCurrent.vue'
+import AskMeAnythingHelp from './src/components/game/AskMeAnythingHelp.vue'
+import Avatar from './src/components/game/Avatar.vue'
+import CarryCapacityIndicator from './src/components/game/CarryCapacityIndicator.vue'
+import Changelog from './src/components/game/Changelog.vue'
+import CombatMoveDetails from './src/components/game/CombatMoveDetails.vue'
+import CombatMoves from './src/components/game/CombatMoves.vue'
+import Controls from './src/components/game/Controls.vue'
+import CraftDiagram from './src/components/game/CraftDiagram.vue'
+import CraftListItem from './src/components/game/CraftListItem.vue'
+import CreatureDetailsModal from './src/components/game/CreatureDetailsModal.vue'
+import CreatureIcon from './src/components/game/CreatureIcon.vue'
+import CreatureKnowledgeBar from './src/components/game/CreatureKnowledgeBar.vue'
+import CreatureKnowledgeLevelInfo from './src/components/game/CreatureKnowledgeLevelInfo.vue'
+import CreatureName from './src/components/game/CreatureName.vue'
+import CreditsModal from './src/components/game/CreditsModal.vue'
+import CurrencyDisplay from './src/components/game/CurrencyDisplay.vue'
+import DeadIndicator from './src/components/game/DeadIndicator.vue'
+import DisconnectIndicator from './src/components/game/DisconnectIndicator.vue'
+import DisplayImpacts from './src/components/game/DisplayImpacts.vue'
+import EffectIcon from './src/components/game/EffectIcon.vue'
+import Effects from './src/components/game/Effects.vue'
+import EnvironmentOverlay from './src/components/game/EnvironmentOverlay.vue'
+import EssenceIndicator from './src/components/game/EssenceIndicator.vue'
+import EssenceModal from './src/components/game/EssenceModal.vue'
+import GameHelpModule from './src/components/game/GameHelpModule.vue'
+import GameTitle from './src/components/game/GameTitle.vue'
+import HappeningDisplay from './src/components/game/HappeningDisplay.vue'
+import HitText from './src/components/game/HitText.vue'
+import ImpactsSummary from './src/components/game/ImpactsSummary.vue'
+import LanguageIncluded from './src/components/game/LanguageIncluded.vue'
+import LanguagePhrase from './src/components/game/LanguagePhrase.vue'
+import Location from './src/components/game/Location.vue'
+import LocationSpacing from './src/components/game/LocationSpacing.vue'
+import MainEntityKnowledgeLevelInfo from './src/components/game/MainEntityKnowledgeLevelInfo.vue'
+import MilestoneInfo from './src/components/game/MilestoneInfo.vue'
+import MilestoneObjective from './src/components/game/MilestoneObjective.vue'
+import MilestoneTracker from './src/components/game/MilestoneTracker.vue'
+import MusicPlayer from './src/components/game/MusicPlayer.vue'
+import OpenQuestionList from './src/components/game/OpenQuestionList.vue'
+// import OpenQuestionPanel from './src/components/game/OpenQuestionPanel.vue'
+import Operation from './src/components/game/Operation.vue'
+import OperationToolSelector from './src/components/game/OperationToolSelector.vue'
+import OwnEffectByName from './src/components/game/OwnEffectByName.vue'
+import PlanListItem from './src/components/game/PlanListItem.vue'
+import PluginSettings from './src/components/game/PluginSettings.vue'
+import PowerGroup from './src/components/game/PowerGroup.vue'
+import PowerItem from './src/components/game/PowerItem.vue'
+import QuickActions from './src/components/game/QuickActions.vue'
+import ResearchCard from './src/components/game/ResearchCard.vue'
+import ResourceIcon from './src/components/game/ResourceIcon.vue'
+import RichText from './src/components/game/RichText.vue'
+import SkillBar from './src/components/game/SkillBar.vue'
+import SkillInfoDisplay from './src/components/game/SkillInfoDisplay.vue'
+import Status from './src/components/game/Status.vue'
+import StructureIcon from './src/components/game/StructureIcon.vue'
+import StructureSelector from './src/components/game/StructureSelector.vue'
+import Trade from './src/components/game/Trade.vue'
+import TradeSide from './src/components/game/TradeSide.vue'
+import UserSettings from './src/components/game/UserSettings.vue'
+
+export const loadAll = (app) => {
+  app.component('Backgrounds', Backgrounds)
+  app.component('BorderRound', BorderRound)
+  app.component('Container', Container)
+  app.component('DynamicLayout', DynamicLayout)
+  app.component('Horizontal', Horizontal)
+  app.component('HorizontalCenter', HorizontalCenter)
+  app.component('HorizontalFill', HorizontalFill)
+  app.component('HorizontalWrap', HorizontalWrap)
+  app.component('PageLayout', PageLayout)
+  app.component('Spaced', Spaced)
+  app.component('Vertical', Vertical)
+
+  app.component('AppStyles', AppStyles)
+  app.component('AutoSizer', AutoSizer)
+  app.component('Button', Button)
+  app.component('Checkbox', Checkbox)
+  app.component('CloseButton', CloseButton)
+  app.component('Collapsible', Collapsible)
+  app.component('Countdown', Countdown)
+  app.component('Description', Description)
+  app.component('EditButton', EditButton)
+  app.component('ExplanationIndicator', ExplanationIndicator)
+  app.component('Header', Header)
+  app.component('Help', Help)
+  app.component('Icon', Icon)
+  app.component('Input', Input)
+  app.component('LabeledValue', LabeledValue)
+  app.component('ListItem', ListItem)
+  app.component('LoadingPlaceholder', LoadingPlaceholder)
+  app.component('Modal', Modal)
+  app.component('OnlyIfVisible', OnlyIfVisible)
+  app.component('OptionSelector', OptionSelector)
+  app.component('PluginInjection', PluginInjection)
+  app.component('Plused', Plused)
+  app.component('ProgressBar', ProgressBar)
+  app.component('Radio', Radio)
+  app.component('ReportButton', ReportButton)
+  app.component('Scrollbar', Scrollbar)
+  app.component('Select', Select)
+  app.component('Slider', Slider)
+  app.component('Spinner', Spinner)
+  app.component('StarRating', StarRating)
+  app.component('SubLabeledValue', SubLabeledValue)
+  app.component('TextArea', TextArea)
+  app.component('ToastNotifications', ToastNotifications)
+  app.component('TopZIndex', TopZIndex)
+  app.component('ViewButton', ViewButton)
+
+  app.component('Actions', Actions)
+  app.component('APBar', APBar)
+  app.component('APBarCurrent', APBarCurrent)
+  app.component('AskMeAnythingHelp', AskMeAnythingHelp)
+  app.component('Avatar', Avatar)
+  app.component('CarryCapacityIndicator', CarryCapacityIndicator)
+  app.component('Changelog', Changelog)
+  app.component('CombatMoveDetails', CombatMoveDetails)
+  app.component('CombatMoves', CombatMoves)
+  app.component('Controls', Controls)
+  app.component('CraftDiagram', CraftDiagram)
+  app.component('CraftListItem', CraftListItem)
+  app.component('CreatureDetailsModal', CreatureDetailsModal)
+  app.component('CreatureIcon', CreatureIcon)
+  app.component('CreatureKnowledgeBar', CreatureKnowledgeBar)
+  app.component('CreatureKnowledgeLevelInfo', CreatureKnowledgeLevelInfo)
+  app.component('CreatureName', CreatureName)
+  app.component('CreditsModal', CreditsModal)
+  app.component('CurrencyDisplay', CurrencyDisplay)
+  app.component('DeadIndicator', DeadIndicator)
+  app.component('DisconnectIndicator', DisconnectIndicator)
+  app.component('DisplayImpacts', DisplayImpacts)
+  app.component('EffectIcon', EffectIcon)
+  app.component('Effects', Effects)
+  app.component('EnvironmentOverlay', EnvironmentOverlay)
+  app.component('EssenceIndicator', EssenceIndicator)
+  app.component('EssenceModal', EssenceModal)
+  app.component('GameHelpModule', GameHelpModule)
+  app.component('GameTitle', GameTitle)
+  app.component('HappeningDisplay', HappeningDisplay)
+  app.component('HitText', HitText)
+  app.component('ImpactsSummary', ImpactsSummary)
+  app.component('LanguageIncluded', LanguageIncluded)
+  app.component('LanguagePhrase', LanguagePhrase)
+  app.component('Location', Location)
+  app.component('LocationSpacing', LocationSpacing)
+  app.component('MainEntityKnowledgeLevelInfo', MainEntityKnowledgeLevelInfo)
+  app.component('MilestoneInfo', MilestoneInfo)
+  app.component('MilestoneObjective', MilestoneObjective)
+  app.component('MilestoneTracker', MilestoneTracker)
+  app.component('MusicPlayer', MusicPlayer)
+  app.component('OpenQuestionList', OpenQuestionList)
+  // app.component('OpenQuestionPanel', OpenQuestionPanel)
+  app.component('Operation', Operation)
+  app.component('OperationToolSelector', OperationToolSelector)
+  app.component('OwnEffectByName', OwnEffectByName)
+  app.component('PlanListItem', PlanListItem)
+  app.component('PluginSettings', PluginSettings)
+  app.component('PowerGroup', PowerGroup)
+  app.component('PowerItem', PowerItem)
+  app.component('QuickActions', QuickActions)
+  app.component('ResearchCard', ResearchCard)
+  app.component('ResourceIcon', ResourceIcon)
+  app.component('RichText', RichText)
+  app.component('SkillBar', SkillBar)
+  app.component('SkillInfoDisplay', SkillInfoDisplay)
+  app.component('Status', Status)
+  app.component('StructureIcon', StructureIcon)
+  app.component('StructureSelector', StructureSelector)
+  app.component('Trade', Trade)
+  app.component('TradeSide', TradeSide)
+  app.component('UserSettings', UserSettings)
 }
-
-const requireComponents = require.context("./src/components/", true, /\.vue$/);
-requireComponents.keys().forEach((fileName) => {
-  const componentName = fileName
-    .split("/")
-    .pop()
-    .replace(/\.\w+$/, "");
-  if (components[componentName]) {
-    throw new Error(
-      `Duplicate component definition found: ${componentName}, found in:\n${fileName}\n${components[componentName]}`
-    );
-  }
-  components[componentName] = fileName;
-  Vue.component(componentName, requireComponents(fileName).default);
-});
-
-const pluginSources = require.context("./src/plugins/", true, /\.vue$/);
-
-const pluginDefinitions = require.context(
-  "./src/plugins/",
-  true,
-  /\/spec\.json$/
-);
-pluginDefinitions.keys().forEach((fileName) => {
-  const spec = pluginDefinitions(fileName);
-  const pluginPath = fileName.replace("spec.json", "");
-
-  spec.id = pluginPath.replace(/^.\//, "").replace(/\/$/, "");
-  if (allPlugins[spec.name]) {
-    spec.error = `Conflicting plugin name: ${spec.name}, found in:\n${fileName}`;
-  } else {
-    spec.componentNames = {};
-    Object.keys(spec.additions).forEach((slot) => {
-      const file = spec.additions[slot];
-      try {
-        const componentName = `PLUGIN_${spec.id}___${slot}`.replace(/-/g, "_");
-        spec.componentNames[slot] = componentName;
-        Vue.component(
-          componentName,
-          pluginSources(`${pluginPath}${file}`).default
-        );
-      } catch (e) {
-        spec.error = `Error in ${file}: ${e.toString()}`;
-        console.warn(`Failed to load plugin ${fileName}: ${spec.error}`);
-      }
-    });
-  }
-
-  allPlugins[spec.name] = spec;
-});
-
-global.allPlugins = Object.values(allPlugins);
