@@ -1,64 +1,85 @@
-<template>
-  <div id="app">
-    <Container
-      v-if="invalidClientVersion"
-      :top="false"
-      :left="false"
-      :right="false"
-      :bottom="false"
-    >
-      <Modal dialog>
-        <template v-slot:title> Game client out of date </template>
-        <template v-slot:contents>
-          <Spaced>
-            <Description>
-              <div>It appears that this client is out of date.</div>
-              <div>Please refresh the page to update your client.</div>
-              <div>
-                If that doesn't work please let us know on the
-                <a href="https://discord.gg/XExbewT5GQ" target="_blank"
-                  >Discord</a
-                >.
-              </div>
-              <hr />
-              <div>Debug info: {{ startupId }} vs {{ currentRev }}</div>
-            </Description>
-          </Spaced>
-        </template>
-      </Modal>
-    </Container>
-    <router-view v-else />
-    <ToastNotifications />
-  </div>
-</template>
-
-<script>
-export default {
-  subscriptions() {
-    const clientStartupId = GameService.getClientStartupId();
-    const startupIdStream = GameService.getStartupIdStream();
-    const skipIntegrity = this?.$route?.query?.skipIntegrityCheck !== undefined;
-    return {
-      startupId: startupIdStream.tap((startupId) => {
-        console.log("Startup ID:", startupId, skipIntegrity);
-        console.log("Client ver:", clientStartupId);
-      }),
-      currentRev: Rx.Observable.of(clientStartupId),
-      invalidClientVersion: startupIdStream.map(
-        (startupId) =>
-          !skipIntegrity &&
-          `${startupId}` !== "dev" &&
-          `${startupId}` !== `${clientStartupId}`
-      ),
-    };
-  },
-
-  created() {},
-};
+<script setup>
+import { RouterLink, RouterView } from 'vue-router'
+import HelloWorld from './components/HelloWorld.vue'
 </script>
 
-<style scoped lang="scss">
-.invalid-version-wrapper {
-  padding: 25%;
+<template>
+  <header>
+    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+
+    <div class="wrapper">
+      <HelloWorld msg="You did it!" />
+
+      <nav>
+        <RouterLink to="/">Home</RouterLink>
+        <RouterLink to="/about">About</RouterLink>
+      </nav>
+    </div>
+  </header>
+
+  <RouterView />
+</template>
+
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
+}
+
+.logo {
+  display: block;
+  margin: 0 auto 2rem;
+}
+
+nav {
+  width: 100%;
+  font-size: 12px;
+  text-align: center;
+  margin-top: 2rem;
+}
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
+}
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+@media (min-width: 1024px) {
+  header {
+    display: flex;
+    place-items: center;
+    padding-right: calc(var(--section-gap) / 2);
+  }
+
+  .logo {
+    margin: 0 2rem 0 0;
+  }
+
+  header .wrapper {
+    display: flex;
+    place-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  nav {
+    text-align: left;
+    margin-left: -1rem;
+    font-size: 1rem;
+
+    padding: 1rem 0;
+    margin-top: 1rem;
+  }
 }
 </style>
