@@ -22,18 +22,13 @@
           class="action-button"
           :class="action.actionId"
         >
-          <RichText
-            class="nowrap"
-            :value="action.label ? action.label : 'Missing label'"
-          />
+          <RichText class="nowrap" :value="action.label ? action.label : 'Missing label'" />
         </Button>
       </div>
     </component>
     <Modal dialog large v-if="currentAction" @close="closeActionDialog()">
       <template v-slot:title>
-        <RichText
-          :value="currentAction.label ? currentAction.label : 'Missing label'"
-        />
+        <RichText :value="currentAction.label ? currentAction.label : 'Missing label'" />
       </template>
       <template v-slot:contents>
         <Vertical class="action-dialog">
@@ -46,11 +41,8 @@
             v-bind="currentAction.extra"
           />
           <div>
-            <LoadingPlaceholder
-              :size="3.5"
-              v-if="actionPoints && calculatingAp"
-            />
-            <APBar v-else-if="actionPoints" />
+            <LoadingPlaceholder :size="3.5" v-if="actionPoints && calculatingAp" />
+            <APBarCurrent v-else-if="actionPoints" />
           </div>
           <div
             v-for="parameter in currentAction.parameters"
@@ -105,11 +97,7 @@
             </div>
             <div v-else>Missing implementation: {{ parameter }}</div>
           </div>
-          <Button
-            @click="performAction(currentAction)"
-            ref="submit"
-            :processing="performing"
-          >
+          <Button @click="performAction(currentAction)" ref="submit" :processing="performing">
             {{ confirmLabel }}
           </Button>
         </Vertical>
@@ -119,19 +107,16 @@
 </template>
 
 <script>
-import exclamationIcon from "../../assets/ui/cartoon/icons/exclamation.png";
-import LoadingPlaceholder from "../interface/LoadingPlaceholder";
-import StructureSelector from "./StructureSelector";
+import exclamationIcon from '../../assets/ui/cartoon/icons/exclamation.png'
 
 export default {
-  components: { StructureSelector, LoadingPlaceholder },
   props: {
     target: Object,
     vertical: {
       type: Boolean,
     },
     confirmLabel: {
-      default: "Confirm",
+      default: 'Confirm',
     },
     noWrap: {
       type: Boolean,
@@ -154,16 +139,16 @@ export default {
     currentAction: null,
     performing: null,
     holdingsFilter: (building) => {
-      return building.own;
+      return building.own
     },
   }),
   subscriptions() {
     return {
       actionPoints: Rx.combineLatest(
-        this.$stream("currentAction"),
-        this.$stream("parameters"),
-        this.$stream("target"),
-        GameService.getRootEntityStream()
+        this.$stream('currentAction'),
+        this.$stream('parameters'),
+        this.$stream('target'),
+        GameService.getRootEntityStream(),
       )
         .distinctUntilChanged(null, JSON.stringify)
         .filter(([currentAction]) => !!currentAction)
@@ -175,52 +160,50 @@ export default {
               targetId: target.id,
               actionId: currentAction.actionId,
               parameters,
-            })
-          )
+            }),
+          ),
         )
         .filter((result) => result && result.ok)
         .map((result) => result.ap)
         .tap((ap) => {
-          this.calculatingAp = false;
-          this.setConsideredAP(ap);
+          this.calculatingAp = false
+          this.setConsideredAP(ap)
         }),
-    };
+    }
   },
 
   computed: {
     filteredActions() {
       if (!this.target || !this.target.actions) {
-        return [];
+        return []
       }
       if (!this.actionId) {
-        return this.target.actions;
+        return this.target.actions
       }
-      const onlyAction = this.target.actions.find(
-        (a) => a.actionId === this.actionId
-      );
+      const onlyAction = this.target.actions.find((a) => a.actionId === this.actionId)
       if (!onlyAction) {
-        return [];
+        return []
       }
-      return [onlyAction];
+      return [onlyAction]
     },
 
     inventoryWeightChange() {
       if (!this.currentAction || !this.target) {
-        return 0;
+        return 0
       }
-      let multiplier;
+      let multiplier
       switch (this.currentAction.actionId) {
-        case "pickup":
-        case "loot":
-          multiplier = 1;
-          break;
-        case "drop":
-          multiplier = -1;
-          break;
+        case 'pickup':
+        case 'loot':
+          multiplier = 1
+          break
+        case 'drop':
+          multiplier = -1
+          break
         default:
-          return 0;
+          return 0
       }
-      return this.target.unitWeight * this.parameters.amount * multiplier;
+      return this.target.unitWeight * this.parameters.amount * multiplier
     },
   },
 
@@ -229,89 +212,87 @@ export default {
       this.parameters = {
         ...this.parameters,
         [parameter.paramId]: structure.id,
-      };
+      }
     },
     updateItemValue(parameter, value) {
       this.parameters = {
         ...this.parameters,
         [parameter.paramId]: value.id,
-      };
+      }
     },
     updateParameterValue(parameter, value) {
       this.parameters = {
         ...this.parameters,
         [parameter.paramId]: value,
-      };
+      }
     },
 
     getMax(parameter) {
-      if (parameter.paramId === "amount" && this.selectedItem) {
-        return this.selectedItem.amount;
+      if (parameter.paramId === 'amount' && this.selectedItem) {
+        return this.selectedItem.amount
       }
-      return parameter.max;
+      return parameter.max
     },
     startAction(action) {
       if (!action.parameters?.length && !action.actionPoints && !action.extra) {
-        this.performAction(action);
-        return;
+        this.performAction(action)
+        return
       }
-      this.currentAction = action;
+      this.currentAction = action
       this.parameters = (action.parameters || []).reduce(
         (acc, param) => ({
           [param.paramId]: param.default,
           ...acc,
         }),
-        {}
-      );
+        {},
+      )
     },
     setConsideredAP(value) {
       if (this.lastConsideredAP !== undefined) {
-        ControlsService.updateConsideredAP(value);
+        ControlsService.updateConsideredAP(value)
       } else {
         ControlsService.getConsideredAPStream()
           .first()
           .subscribe((current) => {
-            this.lastConsideredAP = current;
-            ControlsService.updateConsideredAP(value);
-          });
+            this.lastConsideredAP = current
+            ControlsService.updateConsideredAP(value)
+          })
       }
     },
     performAction(action) {
       const parameters = {
         ...this.parameters,
         ...this.parameterValues,
-      };
-      if (this.customActionHandler) {
-        return this.customActionHandler(action, parameters);
       }
-      this.performingActionId = action.actionId;
-      this.performing = GameService.performAction(
-        this.target,
-        action,
-        parameters
-      ).then((response) => {
-        this.performingActionId = null;
-        if (response && response.ok) {
-          this.closeActionDialog();
-          this.$emit("action", { action, result: response.result });
-          if (action.actionId === "trade") {
-            ControlsService.triggerControlEvent("openPanel-trade");
+      if (this.customActionHandler) {
+        return this.customActionHandler(action, parameters)
+      }
+      this.performingActionId = action.actionId
+      this.performing = GameService.performAction(this.target, action, parameters).then(
+        (response) => {
+          this.performingActionId = null
+          if (response && response.ok) {
+            this.closeActionDialog()
+            this.$emit('action', { action, result: response.result })
+            if (action.actionId === 'trade') {
+              ControlsService.triggerControlEvent('openPanel-trade')
+            }
           }
-        }
-      });
+        },
+      )
     },
     closeActionDialog() {
-      this.currentAction = null;
+      this.currentAction = null
       if (this.lastConsideredAP !== undefined) {
-        this.setConsideredAP(this.lastConsideredAP);
+        this.setConsideredAP(this.lastConsideredAP)
       }
     },
   },
-};
+}
 </script>
 
 <style lang="scss">
-@import "../../utils.scss";
+@import '../../utils.scss';
 
 .action-button-wrapper.disabled {
   pointer-events: none;
