@@ -8,11 +8,7 @@
           class="report-name-button"
           v-if="creature.avatar"
           title="Report name"
-          :description="
-            'Report player name <em>' +
-            creature.name +
-            '</em> as inappropriate.'
-          "
+          :description="'Report player name <em>' + creature.name + '</em> as inappropriate.'"
           type="name"
           :refId="creature.id"
         />
@@ -25,11 +21,7 @@
           <Vertical>
             <Vertical>
               <div class="creature-icon-wrapper">
-                <Avatar
-                  v-if="creature.avatar"
-                  :creature="creature"
-                  size="large"
-                />
+                <Avatar v-if="creature.avatar" :creature="creature" size="large" />
                 <CreatureIcon v-else :creature="creature" size="large" />
               </div>
               <div v-if="creature.nextMove">
@@ -65,10 +57,7 @@
                           }"
                         >
                           {{ creature.operationInfo.name }}
-                          <div
-                            v-if="creature.operationInfo.subtext"
-                            class="subtext"
-                          >
+                          <div v-if="creature.operationInfo.subtext" class="subtext">
                             {{ creature.operationInfo.subtext }}
                           </div>
                         </div>
@@ -92,10 +81,7 @@
               <div class="creature-effects">
                 <Effects :effects="creature.effects" :filter="combatEffects" />
                 <Effects :effects="creature.tracks" />
-                <Effects
-                  :effects="creature.effects"
-                  :filter="nonCombatEffects"
-                />
+                <Effects :effects="creature.effects" :filter="nonCombatEffects" />
               </div>
             </Vertical>
             <div v-if="creature.description">
@@ -108,10 +94,7 @@
           <div v-if="creature.hostile">
             <LoadingPlaceholder v-if="!mobInfo" />
             <div v-else>
-              <CreatureKnowledgeLevelInfo
-                :creature="creature"
-                :mobInfo="mobInfo"
-              />
+              <CreatureKnowledgeLevelInfo :creature="creature" :mobInfo="mobInfo" />
             </div>
           </div>
           <div v-if="creature.id === mainEntity.id">
@@ -126,11 +109,7 @@
 </template>
 
 <script>
-import Horizontal from "../layouts/Horizontal";
-import NextMoveIndicator from "./indicators/NextMoveIndicator";
-import Vertical from "../layouts/Vertical";
 export default {
-  components: { Vertical, NextMoveIndicator, Horizontal },
   props: {
     creatureId: {},
   },
@@ -141,36 +120,33 @@ export default {
   }),
 
   subscriptions() {
-    const creatureStream = this.$stream("creatureId").switchMap((creatureId) =>
+    const creatureStream = this.$stream('creatureId').switchMap((creatureId) =>
       creatureId
         ? GameService.getEntityStream(creatureId, ENTITY_VARIANTS.DETAILS)
-        : Rx.Observable.of(null)
-    );
+        : Rx.Observable.of(null),
+    )
     return {
       mainEntity: GameService.getRootEntityStream(),
       creature: creatureStream,
       targetText: Rx.combineLatest(
-        creatureStream.pluck("nextMove", "targetId"),
-        GameService.getRootEntityStream()
+        creatureStream.pluck('nextMove', 'targetId'),
+        GameService.getRootEntityStream(),
       ).switchMap(([creatureId, mainEntity]) =>
         creatureId === mainEntity.id
-          ? Rx.Observable.of("you")
-          : GameService.getEntityStream(
-              creatureId,
-              ENTITY_VARIANTS.DETAILS
-            ).map((c) => c?.name)
+          ? Rx.Observable.of('you')
+          : GameService.getEntityStream(creatureId, ENTITY_VARIANTS.DETAILS).map((c) => c?.name),
       ),
       mobInfo: creatureStream
         .filter((creature) => !!creature && creature.hostile)
-        .pluck("publicId")
+        .pluck('publicId')
         .switchMap((publicId) =>
           Rx.fromPromise(
             GameService.request(REQUEST_CODES.MOB_INFO, {
               publicId,
-            })
-          )
+            }),
+          ),
         ),
-    };
+    }
   },
 
   methods: {
@@ -181,17 +157,17 @@ export default {
         },
         {
           actionId: operationInfo.action.actionId,
-        }
+        },
       ).then(() => {
-        this.$emit("close");
-      });
+        this.$emit('close')
+      })
     },
   },
-};
+}
 </script>
 
 <style scoped lang="scss">
-@import "../../utils.scss";
+@import '../../utils.scss';
 
 .creature-icon-wrapper {
   margin: 0 auto;
