@@ -2,18 +2,15 @@
   <Vertical>
     <Header>Available plugins</Header>
     <LoadingPlaceholder v-if="!pluginsEnabled" />
-    <Container v-else v-for="plugin in workingPlugins">
+    <Container v-else v-for="plugin in workingPlugins" :key="plugin.name">
       <Spaced>
         <Horizontal>
           <div class="flex-grow">
             <div class="name">{{ plugin.name }}</div>
             <div class="author">by {{ plugin.author }}</div>
           </div>
-          <Button
-            v-if="plugin.settings.length && true"
-            @click="toggleConfig(plugin)"
-          >
-            {{ showSettings[plugin.id] ? "Hide" : "Show" }} Config
+          <Button v-if="plugin.settings.length && true" @click="toggleConfig(plugin)">
+            {{ showSettings[plugin.id] ? 'Hide' : 'Show' }} Config
           </Button>
           <Checkbox
             :value="pluginsEnabled.some((p) => p.id === plugin.id)"
@@ -43,14 +40,14 @@
         <template v-slot:header> Disabled plugins </template>
         <template v-slot:content>
           <Vertical>
-            <Container v-for="plugin in disabledPlugins">
+            <Container v-for="plugin in disabledPlugins" :key="plugin.name">
               <Spaced>
                 <Horizontal>
                   <div class="flex-grow">
                     <div class="name">{{ plugin.name }}</div>
                     <div class="author">by {{ plugin.author }}</div>
                   </div>
-                  <div>Reason: {{ plugin.error ? "Error" : "Disabled" }}</div>
+                  <div>Reason: {{ plugin.error ? 'Error' : 'Disabled' }}</div>
                 </Horizontal>
               </Spaced>
             </Container>
@@ -62,11 +59,7 @@
 </template>
 
 <script>
-import Vertical from "../layouts/Vertical";
-import HorizontalCenter from "../layouts/HorizontalCenter";
-import LoadingPlaceholder from "../interface/LoadingPlaceholder";
 export default {
-  components: { LoadingPlaceholder, HorizontalCenter, Vertical },
   data: () => ({
     showSettings: {},
     settingsValues: {},
@@ -78,60 +71,53 @@ export default {
       workingPlugins: PluginService.getWorkingPluginsStream(),
       disabledPlugins: PluginService.getDisabledPluginsStream(),
       pluginSettings: PluginService.getPluginSettingsStream(),
-    };
+    }
   },
 
   methods: {
     toggleConfig(plugin) {
-      const show = !this.showSettings[plugin.id];
-      this.$set(this.showSettings, plugin.id, show);
+      const show = !this.showSettings[plugin.id]
+      this.$set(this.showSettings, plugin.id, show)
       if (show) {
-        this.$set(
-          this.settingsValues,
-          plugin.id,
-          this.settingsValues[plugin.id] || {}
-        );
+        this.$set(this.settingsValues, plugin.id, this.settingsValues[plugin.id] || {})
         plugin.settings.forEach((setting) => {
-          let defaultValue;
+          let defaultValue
           switch (setting.type) {
-            case "string":
-              defaultValue = "";
-              break;
+            case 'string':
+              defaultValue = ''
+              break
           }
-          const currentValue = this.pluginSettings?.[plugin.id]?.[setting.id];
+          const currentValue = this.pluginSettings?.[plugin.id]?.[setting.id]
           this.$set(
             this.settingsValues[plugin.id],
             setting.id,
-            currentValue === undefined ? defaultValue : currentValue
-          );
-        });
+            currentValue === undefined ? defaultValue : currentValue,
+          )
+        })
       }
     },
 
     async togglePlugin(pluginId, value) {
       if (Boolean(this.pluginSettings?.[pluginId]?.enabled) === value) {
-        return;
+        return
       }
       try {
-        await PluginService.togglePlugin(pluginId, value);
+        await PluginService.togglePlugin(pluginId, value)
       } catch (e) {
-        ToastError("Unexpected error");
+        ToastError('Unexpected error')
       }
     },
 
     async saveSettings(plugin) {
       try {
-        await PluginService.updatePluginSettings(
-          plugin.id,
-          this.settingsValues[plugin.id]
-        );
-        this.toggleConfig(plugin);
+        await PluginService.updatePluginSettings(plugin.id, this.settingsValues[plugin.id])
+        this.toggleConfig(plugin)
       } catch (e) {
-        ToastError("Unexpected error");
+        ToastError('Unexpected error')
       }
     },
   },
-};
+}
 </script>
 
 <style scoped lang="scss">
