@@ -23,10 +23,7 @@
             {{ CHANCE_LABEL[accidentChance] }}
           </span>
           <span v-if="accidentChance"> / </span>
-          <span
-            v-if="accidentChance"
-            :class="'rate rate-color-' + (5 - accidentSeverity)"
-          >
+          <span v-if="accidentChance" :class="'rate rate-color-' + (5 - accidentSeverity)">
             {{ SEVERITY_LABEL[accidentSeverity] }}
           </span>
         </span>
@@ -48,8 +45,7 @@
       <template v-slot:title> Speed </template>
       <template v-slot:contents>
         <Description prominent>
-          Speed determines how many Action Points are used to perform an
-          action.<br />
+          Speed determines how many Action Points are used to perform an action.<br />
           Having higher speed means you spend fewer Action Points.<br />
         </Description>
         <div v-if="speedModifiers">
@@ -57,11 +53,7 @@
           <Description prominent>
             For your current action the speed is determined based on:
           </Description>
-          <LabeledValue
-            v-for="(value, mod) in speedModifiers"
-            :key="mod"
-            :label="mod"
-          >
+          <LabeledValue v-for="(value, mod) in speedModifiers" :key="mod" :label="mod">
             {{ value }}%
             <div v-if="mod === 'Efficiency'" class="stats-icons">
               <Effects
@@ -79,10 +71,7 @@
                 :size="2.5"
               />
             </div>
-            <div
-              v-if="mod === 'Attributes' && !!skillDetails"
-              class="stats-icons"
-            >
+            <div v-if="mod === 'Attributes' && !!skillDetails" class="stats-icons">
               <StatIcon
                 v-for="stat in relatedStatsSorted"
                 :key="stat"
@@ -114,8 +103,7 @@
           </span>
         </LabeledValue>
         <Description prominent>
-          Accident chance depends on the difficulty of the task and your
-          skill.<br />
+          Accident chance depends on the difficulty of the task and your skill.<br />
         </Description>
         <hr />
         <LabeledValue label="Accident Severity">
@@ -124,10 +112,9 @@
           </span>
         </LabeledValue>
         <Description prominent>
-          Accident severity determines how big an injury you may sustain if an
-          accident happens.<br />
-          It depends on the difficulty of the task, your skill, base Action
-          Point cost and other factors.
+          Accident severity determines how big an injury you may sustain if an accident happens.<br />
+          It depends on the difficulty of the task, your skill, base Action Point cost and other
+          factors.
         </Description>
       </template>
     </Modal>
@@ -145,8 +132,7 @@
           </span>
         </LabeledValue>
         <Description prominent>
-          Success chance depends on the difficulty of the task, your skill and
-          other factors.
+          Success chance depends on the difficulty of the task, your skill and other factors.
         </Description>
       </template>
     </Modal>
@@ -154,26 +140,24 @@
 </template>
 
 <script>
-import Horizontal from "../layouts/Horizontal";
 const CHANCE_LABEL = {
-  0: "Impossible",
-  1: "Unlikely",
-  2: "Doubtful",
-  3: "Fair",
-  4: "Likely",
-  5: "Certain",
-};
+  0: 'Impossible',
+  1: 'Unlikely',
+  2: 'Doubtful',
+  3: 'Fair',
+  4: 'Likely',
+  5: 'Certain',
+}
 const SEVERITY_LABEL = {
-  0: "None",
-  1: "Trivial",
-  2: "Minor",
-  3: "Serious",
-  4: "Severe",
-  5: "Grievous",
-};
+  0: 'None',
+  1: 'Trivial',
+  2: 'Minor',
+  3: 'Serious',
+  4: 'Severe',
+  5: 'Grievous',
+}
 
 export default {
-  components: { Horizontal },
   props: {
     operation: {},
     skillData: {},
@@ -181,8 +165,7 @@ export default {
 
   data: () => ({
     explain: null,
-    efficiencyEffectFilter: (effect) =>
-      effect?.impacts?.some((i) => i.name === "Efficiency"),
+    efficiencyEffectFilter: (effect) => effect?.impacts?.some((i) => i.name === 'Efficiency'),
     CHANCE_LABEL,
     SEVERITY_LABEL,
   }),
@@ -190,73 +173,73 @@ export default {
   subscriptions() {
     return {
       mainEntity: GameService.getRootEntityStream(),
-      skillDetails: this.$stream("operation")
-        .pluck("context", "skillInfo", "skill")
+      skillDetails: this.$stream('operation')
+        .pluck('context', 'skillInfo', 'skill')
         .switchMap((skillName) =>
-          GameService.getInfoStream("SKILLS", { skillName: skillName }, true)
+          GameService.getInfoStream('SKILLS', { skillName: skillName }, true),
         ),
-    };
+    }
   },
 
   computed: {
     relatedStatsSorted() {
       if (!this.skillDetails || !this.skillDetails.relatedStats) {
-        return [];
+        return []
       }
-      return [...this.skillDetails.relatedStats].sort();
+      return [...this.skillDetails.relatedStats].sort()
     },
     speedClass() {
       switch (true) {
         case this.finalSpeed < 25:
-          return "rate-color-0";
+          return 'rate-color-0'
         case this.finalSpeed < 50:
-          return "rate-color-1";
+          return 'rate-color-1'
         case this.finalSpeed < 75:
-          return "rate-color-2";
+          return 'rate-color-2'
         case this.finalSpeed < 100:
-          return "rate-color-3";
+          return 'rate-color-3'
         case this.finalSpeed < 120:
-          return "rate-color-4";
+          return 'rate-color-4'
         default:
-          return "rate-color-5";
+          return 'rate-color-5'
       }
     },
     internalSkillData() {
-      return this.skillData || this.operation?.context?.skillInfo;
+      return this.skillData || this.operation?.context?.skillInfo
     },
     finalSpeed() {
-      return this.internalSkillData?.finalSpeed;
+      return this.internalSkillData?.finalSpeed
     },
     baseSpeed() {
-      return this.internalSkillData?.baseSpeed;
+      return this.internalSkillData?.baseSpeed
     },
     speedModifiers() {
-      return this.internalSkillData?.speedModifiers;
+      return this.internalSkillData?.speedModifiers
     },
     skill() {
-      return this.internalSkillData?.skill;
+      return this.internalSkillData?.skill
     },
     successChance() {
-      return this.internalSkillData?.successChance;
+      return this.internalSkillData?.successChance
     },
     accidentChance() {
-      return this.internalSkillData?.accidentChance;
+      return this.internalSkillData?.accidentChance
     },
     accidentSeverity() {
-      return this.internalSkillData?.accidentSeverity;
+      return this.internalSkillData?.accidentSeverity
     },
     expGainMultiplier() {
-      return this.internalSkillData?.skillGainMult;
+      return this.internalSkillData?.skillGainMult
     },
     skillLevel() {
-      return this.internalSkillData?.skillLevel;
+      return this.internalSkillData?.skillLevel
     },
   },
-};
+}
 </script>
 
 <style scoped lang="scss">
-@import "../../utils.scss";
+@import '../../utils.scss';
 
 .efficiency-effects {
   display: inline-block;
